+++ GeneradorDeExamenes.Web/Controllers/IaController.cs
@@ -1,5 +1,4 @@
 ﻿using GeneradorDeExamanes.Logica.Services;
-using GeneradorDeExamenes.Logica.Services;
 using GeneradorDeExamenes.Models;
 
 using Microsoft.AspNetCore.Mvc;
